backend/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.1
django-cors-headers==3.11.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
importlib-metadata==4.10.0
Markdown==3.3.6
Pillow==9.0.0
psycopg2==2.9.3
pytesseract==0.3.8
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.0.0
zipp==3.7.0
